feat: Se mejoro el codigo, se hizo la vista mas amigable y solucion a leaks de seguridad

+++ src/Modules/Blog/Application/Commands/DeleteCategoryCommand.php
@@ -8,8 +8,14 @@ class DeleteCategoryCommand
 {
     public function __construct(private readonly ICategoryRepository $repository) { }
 
-    public function deleteCategory(array $data): void
+    public function deleteCategory(array $data): bool
     {
+        $category = $this->repository->getCategory($data);
+
+        if ($category->posts()->count() > 0) return false;
+
         $this->repository->delete($data);
+
+        return true;
     }
 }
